+++ src/main/java/com/mandarina/game/main/GameCts.java
@@ -6,9 +6,9 @@ public class GameCts {
 	public final static float SCALE_DEFAULT = 1f;
 	public final static int TILES_IN_WIDTH = 26;
 	public final static int TILES_IN_HEIGHT = 14;
-	public final static double GAME_WIDTH_DEFAULT = GameCts.TILES_DEFAULT_SIZE * GameCts.TILES_IN_WIDTH;
-	public final static double GAME_HEIGHT_DEFAULT = GameCts.TILES_DEFAULT_SIZE * GameCts.TILES_IN_HEIGHT;
+	public final static float GAME_WIDTH_DEFAULT = GameCts.TILES_DEFAULT_SIZE * GameCts.TILES_IN_WIDTH;
+	public final static float GAME_HEIGHT_DEFAULT = GameCts.TILES_DEFAULT_SIZE * GameCts.TILES_IN_HEIGHT;
 	public final static int ANI_SPEED = 25;
-	public final static double GRAVITY_DEFAULT = 0.03f;
+	public final static float GRAVITY_DEFAULT = 0.03f;
 
 }
