+++ src/main/java/com/mandarina/game/objects/ProjectileCts.java
@@ -3,5 +3,5 @@ package com.mandarina.game.objects;
 public class ProjectileCts {
 	public static final int CANNON_BALL_WIDTH_DEFAULT = 15;
 	public static final int CANNON_BALL_HEIGHT_DEFAULT = 15;
-	public static final double SPEED_DEFAULT = 0.75f;
+	public static final float SPEED_DEFAULT = 0.75f;
 }
